# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  utils.PaginatedResponseGeneric:
    properties:
      data:
        description: Menggunakan interface{} untuk mewakili data apa pun
        items: {}
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.PaginationMeta'
      success:
        type: boolean
    type: object
  utils.PaginationMeta:
    properties:
      current_page:
        description: Halaman saat ini.
        type: integer
      per_page:
        description: Jumlah item per halaman yang digunakan (limit).
        type: integer
      total_items:
        description: Total jumlah item di semua halaman.
        type: integer
      total_pages:
        description: Total jumlah halaman yang tersedia.
        type: integer
    type: object
host: localhost:3001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API backend for digital parenting application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Digital Parenting App BE API
  version: "1.0"
paths:
...
  /child/claims:
    get:
      description: Retrieves the history of reward claims made by the logged-in child
        (paginated).
      parameters:
      - description: Filter by status (pending, approved, rejected)
        enum:
        - pending
        - approved
        - rejected
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Claims history retrieved
          schema:
            $ref: '#/definitions/utils.PaginatedResponseGeneric'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get My Reward Claims History
      tags:
      - Child - Points & Rewards
  /child/points:
    get:
      description: Retrieves the current points balance for the logged-in child.
      produces:
      - application/json
      responses:
        "200":
          description: Points balance retrieved
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get My Points Balance
      tags:
      - Child - Points & Rewards
  /child/points/history:
    get:
      description: Retrieves the points transaction history for the logged-in child
        (paginated).
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Points history retrieved
          schema:
            $ref: '#/definitions/utils.PaginatedResponseGeneric'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get My Points History
      tags:
      - Child - Points & Rewards
  /child/rewards:
    get:
      description: Retrieves rewards available for the logged-in child to claim.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Available rewards retrieved
          schema:
            $ref: '#/definitions/utils.PaginatedResponseGeneric'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Available Rewards
      tags:
      - Child - Points & Rewards
  /child/rewards/{rewardId}/claim:
    post:
      description: Submits a claim request for a specific reward by the logged-in
        child.
      parameters:
      - description: Reward ID to claim
        in: path
        name: rewardId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Reward claim submitted
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
        "400":
          description: Invalid Reward ID
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "402":
          description: Insufficient points" // Payment Required (402) bisa dipakai
            di sini
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Reward not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Claim a Reward
      tags:
      - Child - Points & Rewards
  /child/tasks:
    get:
      description: Retrieves tasks assigned to the logged-in child, optionally filtered
        by status.
      parameters:
      - description: Filter by status (assigned, submitted, approved, rejected)
        enum:
        - assigned
        - submitted
        - approved
        - rejected
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks retrieved
          schema:
            $ref: '#/definitions/utils.PaginatedResponseGeneric'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get My Tasks
      tags:
      - Child - Tasks
  /child/tasks/{userTaskId}/submit:
    patch:
      description: Marks a specific assigned task as 'submitted' by the logged-in
        child.
      parameters:
      - description: UserTask ID (the specific assignment)
        in: path
        name: userTaskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task submitted successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Invalid UserTask ID
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden (Not your task or task not assignable)
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Task assignment not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Submit My Task
      tags:
      - Child - Tasks
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer YOUR_JWT_TOKEN'' into the value field."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"